// generated/src/app/api-models/claim-other-carrier.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimOtherCarrier {
  seqClaimId: number;
  sequenceNo: number;
  seqMembId: number;
  claimFilingInd: string;
  sourceOfPay: string;
  insuranceTypeCode: string;
  payorOrganizationId: string;
  payorClaimOfficeNo: string;
  payorName: string;
  groupNo: string;
  groupName: string;
  ppoHmoInd: string;
  ppoInd: string;
  priorAuthNo: string;
  assignOfBenefits: string;
  patSignatureSource: string;
  patRelToInsured: string;
  insuredIdNo: string;
  insuredLastName: string;
  insuredFirstName: string;
  insuredMi: string;
  insuredGeneration: string;
  insuredSex: string;
  insuredDateOfBirth: Date;
  insuredEmplStatus: string;
  supplementalInsInd: string;
  insuranceLocationId: string;
  payorAddr1: string;
  payorAddr2: string;
  payorCity: string;
  payorState: string;
  payorZip: string;
  disallowedCostCont: number;
  disallowedOther: number;
  allowedAmount: number;
  deductibleAmount: number;
  coinsuranceAmount: number;
  payorAmountPaid: number;
  zeroPaidInd: string;
  adjudicationInd1: string;
  adjudicationInd2: string;
  adjudicationInd3: string;
  champusSpnsrBranch: string;
  champusSpnsrGrade: string;
  champusSpnsrStatus: string;
  insCardEffectDate: Date;
  insCardTermDate: Date;
  balanceDue: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  providerNumber: string;
  patReleaseInd: string;
  payorCountry: string;
  treatmentAuthCode: string;
  insuredEmpName: string;
  insuredEmpLocation: string;
  internalControlNo: string;
  securityCode: string;
  cobCode: string;
  cobPtRelCode: string;
}